requirements.txt
google-auth
google-cloud-dns
pyyaml